+++ examples/snippets/api/subscribe_with_state.go
@@ -7,7 +7,7 @@ import (
 )
 
 func main() {
-	config := pubnub.NewConfig(pubnub.GenerateUUID())
+	config := pubnub.NewConfigWithUserId(UserId(pubnub.GenerateUUID()))
 	config.SubscribeKey = "demo"
 	config.PublishKey = "demo"