+++ examples/snippets/design_patterns/advanced_channel_groups.go
@@ -12,7 +12,7 @@ import (
 var pn *pubnub.PubNub
 
 func init() {
-	config := pubnub.NewConfig(pubnub.GenerateUUID())
+	config := pubnub.NewConfigWithUserId(UserId(pubnub.GenerateUUID()))
 	config.SubscribeKey = "sub-c-b9ab9508-43cf-11e8-9967-869954283fb4"
 	config.PublishKey = "pub-c-1bd448ed-05ba-4dbc-81a5-7d6ff5c6e2bb"