+++ remove_channel_channel_group_request_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 func init() {
-	pnconfig = NewConfig(GenerateUUID())
+	pnconfig = NewConfigWithUserId(UserId(GenerateUUID()))
 
 	pnconfig.PublishKey = "pub_key"
 	pnconfig.SubscribeKey = "sub_key"