+++ tests/e2e/delete_channel_group_test.go
@@ -54,7 +54,7 @@ func TestRemoveChannelGroupSuperCall(t *testing.T) {
 
 	validCharacters := "-_~?#[]@!$&'()+;=`|"
 
-	config.UUID = validCharacters
+	config.SetUserId(pubnub.UserId(validCharacters))
 
 	pn := pubnub.NewPubNub(config)