+++ tests/e2e/helper.go
@@ -9,6 +9,7 @@ import (
 	"math/rand"
 	"net"
 	"net/http"
+	"os"
 	"time"
 
 	pubnub "github.com/pubnub/go/v7"
@@ -42,28 +43,27 @@ func seedRand() {
 
 func init() {
 	seedRand()
-	config = pubnub.NewConfig(pubnub.GenerateUUID())
-	config.PublishKey = "pub-c-3ed95c83-12e6-4cda-9d69-c47ba2abb57e"
-	config.SubscribeKey = "sub-c-26a73b0a-c3f2-11e9-8b24-569e8a5c3af3"
+	config = pubnub.NewConfigWithUserId(pubnub.UserId(pubnub.GenerateUUID()))
+	config.PublishKey = os.Getenv("PUBLISH_KEY")
+	config.SubscribeKey = os.Getenv("SUBSCRIBE_KEY")
 
-	pamConfig = pubnub.NewConfig(pubnub.GenerateUUID())
-	pamConfig.PublishKey = "pub-c-cdea0ef1-c571-4b72-b43f-ff1dc8aa4c5d"
-	pamConfig.SubscribeKey = "sub-c-4757f09c-c3f2-11e9-9d00-8a58a5558306"
-	pamConfig.SecretKey = "sec-c-YTYxNzVjYzctNDY2MS00N2NmLTg2NjYtNGRlNWY1NjMxMDBm"
-
+	pamConfig = pubnub.NewConfigWithUserId(pubnub.UserId(pubnub.GenerateUUID()))
+	pamConfig.PublishKey = os.Getenv("PAM_PUBLISH_KEY")
+	pamConfig.SubscribeKey = os.Getenv("PAM_SUBSCRIBE_KEY")
+	pamConfig.SecretKey = os.Getenv("PAM_SECRET_KEY")
 }
 
 func configCopy() *pubnub.Config {
 	cfg := new(pubnub.Config)
 	*cfg = *config
-	cfg.UUID = pubnub.GenerateUUID()
+	cfg.SetUserId(pubnub.UserId(pubnub.GenerateUUID()))
 	return cfg
 }
 
 func pamConfigCopy() *pubnub.Config {
 	config := new(pubnub.Config)
 	*config = *pamConfig
-	config.UUID = pubnub.GenerateUUID()
+	config.SetUserId(pubnub.UserId(pubnub.GenerateUUID()))
 	return config
 }