+++ tests/e2e/list_all_channel_group_test.go
@@ -53,7 +53,7 @@ func TestListAllChannelGroupSuperCall(t *testing.T) {
 
 	validCharacters := "-_~?#[]@!$&'()+;=`|"
 
-	config.UUID = validCharacters
+	config.SetUserId(pubnub.UserId(validCharacters))
 
 	pn := pubnub.NewPubNub(config)